+++ R/mod_global.R
@@ -85,6 +85,12 @@ mod_global_server <- function(input, output, session, orig_data){
 
   world <- reactive({
     orig_data_aggregate_today() %>%
+      align_country_names_pop() %>%
+      mutate(country_name = Country.Region) %>%
+      get_pop_data() %>%
+      mutate(death_rate_pop = round(10^6*deaths/population, digits = 3)) %>%
+      select(-country_name) %>%
+      align_country_names_pop_reverse() %>%
       arrange(desc(confirmed) )
   })